sched/requirements.txt
adaptdl
future>=0.17.1
kubernetes_asyncio==9.1.0  # 10.0.0 is buggy
prometheus_client>=0.7.1
pymoo>=0.4.0
python-dateutil>=2.7.3
requests>=2.7.0
numpy>=1.19
aiohttp>=3.6.2
jsonpatch>=1.26
